ide: stop truncating a 256-sector PIO transfer to zero
ide_transfer_read_once() guards `count <= 256` and then narrowed the
argument to uint8_t, so a count of exactly 256 arrived as 0.

That is not a no-op at either end.  ide_issue_rw() writes the low byte to
the sector-count register, and ATA defines a written 0 as *256 sectors* --
so the drive began transferring 128 KiB.  But the copy loop was
`for (int i = 0; i < count; i++)` with count == 0, so it ran zero times
and the function returned 0 for success.

It is reachable on an ordinary sequential read: blkdev.c sets
BLKDEV_RA_MAX to 256 and clamps the read-ahead window to exactly that, so
blkdev_prefetch() asks for 256 sectors, is told the read succeeded, and
caches 128 KiB of untouched ra_buf heap as valid file data.  The drive is
left with a full transfer pending and DRQ asserted, so the next command on
that channel starts against a desynchronised device.  The write path was
symmetric: success reported having sent nothing.

count is now uint16_t from ide_transfer_*_once() through to the PIO
helpers, which validate 1..256 and iterate the true count while still
letting the 256-as-0 encoding reach the register (ide_issue_rw already
took uint16_t and writes the low byte).

Not exercised by the default qemu config -- run-networking.sh puts the
disk on ich9-ahci even for -machine pc -- so this is verified by
inspection plus a no-regression boot to sdm, not by an IDE transfer.

Refs #391.

+++ sys/drivers/storage/ide/ide.c
@@ -139,7 +139,7 @@ int ide_transfer_read_once(ide_drive_ctx_t *ctx, uint64_t sector,
     }
 
     if (sector < 0x10000000ULL && count <= 256) {
-        return ide_read_sectors(bus, drive, (uint32_t)sector, (uint8_t)count, buffer);
+        return ide_read_sectors(bus, drive, (uint32_t)sector, (uint16_t)count, buffer);
     }
 
     return ide_read_sectors_ext(bus, drive, sector, (uint16_t)count, buffer);
@@ -170,7 +170,7 @@ int ide_transfer_write_once(ide_drive_ctx_t *ctx, uint64_t sector,
     }
 
     if (sector < 0x10000000ULL && count <= 256) {
-        return ide_write_sectors(bus, drive, (uint32_t)sector, (uint8_t)count, buffer);
+        return ide_write_sectors(bus, drive, (uint32_t)sector, (uint16_t)count, buffer);
     }
 
     return ide_write_sectors_ext(bus, drive, sector, (uint16_t)count, buffer);
